backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.10.18
python-multipart==0.0.9
google-generativeai==0.7.2

